Seperated API code from port listener
